requirements.txt
openai
langchain
faiss-cpu
pandas
gradio
